Fix constraint error on playTimerRegulations.bedtimeStartingTime
I'm not sure if this is intended by Nintendo or not, but suddenly it seems bedtimeStartingTime is no longer valid in the schema, at least when sending updates and this isn't set.

Workaround here to to remove that if the hour is set to 0. The integration does not use this anyway